+++ NEWS.md
@@ -1,3 +1,4 @@
 # nanoAFMr 2.1
 
+* fix `summary.AFMdata()` for noImage data types such as force curves
 * data analysis of AFM data files